backend/sports/tests/requirements.txt
requests==2.32.3
pyjwt==2.9.0
aws-lambda-powertools==3.2.0
fastjsonschema==2.21.1
bcrypt==4.1.3
psycopg2-binary==2.9.10
pandas==2.2.3
moto==5.0.24